LOTS OF CHANGES
So, we dropped the byte array, and instead used a packed struct. The
packed struct needs a format string that starts with '=' so that we use
the endianness of the system, but don't allow for byte padding (since we
only want to send 7 bits).

Check the changes for more info